+++ src/BugsnagServiceProvider.php
@@ -186,7 +186,7 @@ class BugsnagServiceProvider extends ServiceProvider
             $client = new Client(new Configuration($config['api_key']), new LaravelResolver($app), $this->getGuzzle($config));
 
             $this->setupCallbacks($client, $app, $config);
-            $this->setupPaths($client, $app->basePath(), $app->path(), isset($config['strip_path']) ? $config['strip_path'] : null, isset($config['project_root']) ? $config['project_root'] : null);
+            $this->setupPaths($client, $app, $config);
 
             $client->setReleaseStage(isset($config['release_stage']) ? $config['release_stage'] : $app->environment());
             $client->setHostname(isset($config['hostname']) ? $config['hostname'] : null);
@@ -328,38 +328,29 @@ class BugsnagServiceProvider extends ServiceProvider
     /**
      * Setup the client paths.
      *
-     * @param \Bugsnag\Client $client
-     * @param string          $base
-     * @param string          $path
-     * @param string|null     $strip
-     * @param string|null     $project
+     * @param \Bugsnag\Client                           $client
+     * @param \Illuminate\Contracts\Container\Container $app
+     * @param array                                     $config
      *
      * @return void
      */
-    protected function setupPaths(Client $client, $base, $path, $strip, $project)
+    protected function setupPaths(Client $client, Container $app, array $config)
     {
-        if ($strip) {
-            $client->setStripPath($strip);
-
-            if (!$project) {
-                $client->setProjectRoot("{$strip}/app");
-            }
-
-            return;
+        if (isset($config['project_root_regex'])) {
+            $client->setProjectRootRegex($config['project_root_regex']);
+        } elseif (isset($config['project_root'])) {
+            $client->setProjectRoot($config['project_root']);
+        } else {
+            $client->setProjectRoot($app->path());
         }
 
-        if ($project) {
-            if ($base && substr($project, 0, strlen($base)) === $base) {
-                $client->setStripPath($base);
-            }
-
-            $client->setProjectRoot($project);
-
-            return;
+        if (isset($config['strip_path_regex'])) {
+            $client->setStripPathRegex($config['strip_path_regex']);
+        } elseif (isset($config['strip_path'])) {
+            $client->setStripPath($config['strip_path']);
+        } else {
+            $client->setStripPath($app->basePath());
         }
-
-        $client->setStripPath($base);
-        $client->setProjectRoot($path);
     }
 
     /**